Blocked because of missing WeaponsFactory

+++ src/main/java/com/example/supergame/model/dto/item/Consumables.java
@@ -4,7 +4,7 @@ package com.example.supergame.model.dto.item;
 public abstract class Consumables extends Item {
 
 
-    Consumables(String name, double price) {
+    Consumables(String name, int price) {
         super(name, price);
     }
 }
