+++ src/main/java/com/example/supergame/model/dto/item/MeleeWeapon.java
@@ -9,7 +9,7 @@ public class MeleeWeapon extends Weapon {
 
     private double damageMultiplier;
 
-    public MeleeWeapon(String name, double price, double damage, double accuracy, WeaponType weaponType, double damageMultiplier) {
+    public MeleeWeapon(String name, int price, double damage, double accuracy, WeaponType weaponType, double damageMultiplier) {
         super(name, price, damage, accuracy, weaponType);
         this.damageMultiplier = damageMultiplier;
     }
